refactor: simplify layouts
