+++ apps/main/src/dao/components/PageAnalytics/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { styled } from 'styled-components'
 import Analytics from '@/dao/components/PageAnalytics/index'
 import { breakpoints } from '@ui/utils'
