+++ apps/main/src/dao/components/PageGauges/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { styled } from 'styled-components'
 import Gauges from '@/dao/components/PageGauges/index'
 import Settings from '@/dao/layout/Settings'
