+++ apps/main/src/dao/components/PageUser/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { styled } from 'styled-components'
 import UserPage from '@/dao/components/PageUser/index'
 import type { UserUrlParams } from '@/dao/types/dao.types'
