+++ apps/main/src/dao/components/PageVeCrv/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect } from 'react'
 import { styled } from 'styled-components'
 import { useAccount } from 'wagmi'
