+++ apps/main/src/dex/components/PageCompensation/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { Contract, Interface } from 'ethers'
 import { useCallback, useEffect, useState } from 'react'
 import { styled } from 'styled-components'
