+++ apps/main/src/dex/components/PageDeployGauge/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { styled } from 'styled-components'
 import { DeployGauge } from '@/dex/components/PageDeployGauge/index'
 import type { NetworkUrlParams } from '@/dex/types/main.types'
