+++ apps/main/src/dex/components/PageIntegrations/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect } from 'react'
 import { styled } from 'styled-components'
 import IntegrationsComp from '@/dex/components/PageIntegrations/index'
