+++ apps/main/src/dex/components/PagePool/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect, useMemo } from 'react'
 import Transfer from '@/dex/components/PagePool/index'
 import { ROUTE } from '@/dex/constants'
