+++ apps/main/src/dex/components/PagePoolList/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { styled } from 'styled-components'
 import PoolList from '@/dex/components/PagePoolList/index'
