+++ apps/main/src/dex/components/PageRouterSwap/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { styled } from 'styled-components'
 import QuickSwap from '@/dex/components/PageRouterSwap/index'
