+++ apps/main/src/llamalend/PageLlamaMarkets/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect } from 'react'
 import { useAccount } from 'wagmi'
 import type { LlamalendServerData } from '@/app/api/llamalend/types'
