+++ apps/main/src/loan/components/PageIntegrations/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect } from 'react'
 import { styled } from 'styled-components'
 import IntegrationsComp from '@/loan/components/PageIntegrations/index'
