+++ apps/main/src/loan/components/PageLoanManage/Page.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { useEffect, useMemo, useState } from 'react'
 import { styled } from 'styled-components'
 import ChartOhlcWrapper from '@/loan/components/ChartOhlcWrapper'
