+++ apps/main/src/loan/components/PagePegKeepers/index.tsx
@@ -1,4 +1,3 @@
-'use client'
 import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
