+++ packages/curve-ui-kit/src/features/layout/store.ts
@@ -1,13 +1,7 @@
-import lodash from 'lodash'
-import { useCallback, useEffect } from 'react'
 import { create } from 'zustand'
 import { devtools } from 'zustand/middleware'
 import { immer } from 'zustand/middleware/immer'
-import { getPageWidthClassName } from '@ui-kit/features/layout/utils'
-import { useUserProfileStore } from '@ui-kit/features/user-profile'
 import type { PageWidthClassName } from './types'
-
-const { delay } = lodash
 
 interface LayoutState {
   windowWidth: number
@@ -79,44 +73,6 @@ const layoutStore = immer<LayoutState & LayoutActions>((set) => ({
       state.isPageVisible = visible
     }),
 }))
-
-export const useLayoutStoreResponsive = () => {
-  const { document } = typeof window === 'undefined' ? {} : window
-  const theme = useUserProfileStore((state) => state.theme)
-  const pageWidth = useLayoutStore((state) => state.pageWidth)
-  const setLayoutWidth = useLayoutStore((state) => state.setLayoutWidth)
-  const setPageVisible = useLayoutStore((state) => state.setPageVisible)
-  const updateShowScrollButton = useLayoutStore((state) => state.updateShowScrollButton)
-
-  const handleResizeListener = useCallback(() => {
-    if (window?.innerWidth) setLayoutWidth(getPageWidthClassName(window.innerWidth))
-  }, [setLayoutWidth])
-
-  useEffect(() => {
-    if (!pageWidth || !document) return
-    document.body.className = `theme-${theme} ${pageWidth}`.replace(/ +(?= )/g, '').trim()
-    document.body.setAttribute('data-theme', theme)
-  }, [document, pageWidth, theme])
-
-  useEffect(() => {
-    if (!window || !document) return
-    const handleScrollListener = () => updateShowScrollButton(window.scrollY)
-    const handleVisibilityChange = () => setPageVisible(!document.hidden)
-
-    handleResizeListener()
-    handleVisibilityChange()
-
-    document.addEventListener('visibilitychange', handleVisibilityChange)
-    window.addEventListener('resize', () => handleResizeListener())
-    window.addEventListener('scroll', () => delay(handleScrollListener, 200))
-
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange)
-      window.removeEventListener('resize', () => handleResizeListener())
-      window.removeEventListener('scroll', () => handleScrollListener())
-    }
-  }, [document, handleResizeListener, setPageVisible, updateShowScrollButton])
-}
 
 export const useLayoutStore =
   process.env.NODE_ENV === 'development' ? create(devtools(layoutStore)) : create(layoutStore)
